+++ x/perp/v2/integration/action/snapshot.go
@@ -20,8 +20,9 @@ type insertReserveSnapshot struct {
 }
 
 func (i insertReserveSnapshot) Do(app *app.NibiruApp, ctx sdk.Context) (sdk.Context, error, bool) {
-	amm := app.PerpKeeperV2.AMMs.GetOr(ctx, i.pair, types.AMM{
+	amm := app.PerpKeeperV2.AMMs.GetOr(ctx, collections.Join(i.pair, uint64(1)), types.AMM{
 		Pair:            i.pair,
+		Version:         uint64(1),
 		BaseReserve:     sdk.ZeroDec(),
 		QuoteReserve:    sdk.ZeroDec(),
 		SqrtDepth:       sdk.ZeroDec(),